// src/frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export interface Article {
    title: string;
    publishDate: Time;
    body: string;
    published: boolean;
    slug: string;
    coverImage: ExternalBlob;
    category: string;
    subtitle?: string;
}
export type Time = bigint;
export interface ArticleSummary {
    title: string;
    publishDate: Time;
    slug: string;
    coverImage: ExternalBlob;
    category: string;
    subtitle?: string;
}
export interface ArticleInput {
    title: string;
    publishDate: Time;
    body: string;
    slug: string;
    coverImage: ExternalBlob;
    category: string;
    subtitle?: string;
}
export interface UserProfile {
    name: string;
}
export enum UserRole {
    admin = "admin",
    user = "user",
    guest = "guest"
}
export interface backendInterface {
    assignCallerUserRole(user: Principal, role: UserRole): Promise<void>;
    checkArticleExists(slug: string): Promise<boolean>;
    createArticle(input: ArticleInput): Promise<void>;
    deleteArticle(slug: string): Promise<void>;
    getArticleBySlug(slug: string): Promise<Article | null>;
    getCallerUserProfile(): Promise<UserProfile | null>;
    getCallerUserRole(): Promise<UserRole>;
    getPendingArticles(): Promise<Array<ArticleSummary>>;
    getPublishedArticles(): Promise<Array<ArticleSummary>>;
    getUserProfile(user: Principal): Promise<UserProfile | null>;
    isCallerAdmin(): Promise<boolean>;
    saveCallerUserProfile(profile: UserProfile): Promise<void>;
    updateArticle(slug: string, input: ArticleInput, published: boolean): Promise<void>;
}